Added more calendar stuff
Courses will only appear on calendar when in the deck (middle column) and selected.
If a card is selected in the left column and dragged into the deck, the card will be unselected and placed in the deck. 
Courses have different colors up to the number of different #calendar-wrapper .occupied CSS classes specified. Some bugs-- if courses overlap, skips over some colors